// src/features/reports/submission/models/reportSubmissionModels.ts
export interface ReportSubmissionPayload {
  id: string;
  referenceNumber: string;
  name: string;
  category: string;
  location: string;
  building: string;
  dateReported: string;
  timeReported: string;
  brand: string;
  color: string;
  identifyingMarks: string;
  description: string;
  contactNumber: string;
  email: string;
  imageUrl: string;
  status: string;
  type: string;
}